Changes score types to decimal
Updates the GroundwaterScore.GWScore and OnSiteScore.OnsiteScoreValue properties to use decimal type instead of string, allowing for more accurate calculations and data representation.

Also, changes Status.CostEstimate to decimal to ensure consistency in data types.

Adds Phone attribute to Phone.Number.

Fixes #798

+++ FMS.Domain/Entities/Status.cs
@@ -108,7 +108,7 @@ namespace FMS.Domain.Entities
         public bool FinancialAssurance { get; set; }
 
         [Display(Name = "Cost Estimate")]
-        public double? CostEstimate { get; set; }
+        public decimal? CostEstimate { get; set; }
 
         [Display(Name = "Cost Estimate Date")]
         public DateOnly? CostEstimateDate { get; set; }
